Andres

+++ src/app/components/list-all-aprendiz/list-all-aprendiz.component.ts
@@ -1,6 +1,9 @@
 import Swal from 'sweetalert2';
 import { Component } from '@angular/core';
 import { AprendizService } from 'src/app/services/aprendiz/aprendiz.service';
+import { ActivatedRoute, Router } from '@angular/router';
+import {FormBuilder} from '@angular/forms';
+import { Ng2SearchPipeModule } from 'ng2-search-filter/src/ng2-filter.module';
 
 @Component({
   selector: 'app-list-all-aprendiz',
@@ -8,11 +11,14 @@ import { AprendizService } from 'src/app/services/aprendiz/aprendiz.service';
   styleUrls: ['./list-all-aprendiz.component.css']
 })
 export class ListAllAprendizComponent {
-
-
+  search:any;
   Aprendiz:any=[];
 
-  constructor(private serviceaprendiz:AprendizService){}
+  filter={
+    contenido:''
+  }
+
+  constructor(private serviceaprendiz:AprendizService,private acrivaterouter: ActivatedRoute,private router:Router, private formBuilder: FormBuilder){}
 
   ngOnInit(): void {
   
@@ -30,34 +36,54 @@ export class ListAllAprendizComponent {
       }
     )
   }
-
   borrar(document: number) {
-    this.serviceaprendiz.deleteAprendiz(document).subscribe(
-      data =>{
-        console.log(data);
-        
-        this.Aprendiz = data;
-        console.log(this.Aprendiz);
-        window.location.reload()
+    Swal.fire({
+      title: 'Desea eliminar este registro? ',
+      showCancelButton: true,
+      confirmButtonText: 'Si',
 
-        // Swal.fire({
-        //   title: '¿desea eliminar este aprendiz?',
-        //   text: "este registro se eliminara para siempre",
-        //   icon: 'warning',
-        //   showCancelButton: true,
-        //   confirmButtonColor: '#3085d6',
-        //   cancelButtonColor: '#d33',
-        //   confirmButtonText: 'si'
-        // }).then((result) => {
-        //   if (result.isConfirmed) {
-       
-        //   }if(result.isDenied){
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.serviceaprendiz.deleteAprendiz(document).subscribe(
+          data =>{
+            console.log(data);
+            
+            this.Aprendiz = data;
+            console.log(this.Aprendiz);
+          }
+        )
+        Swal.fire('El registro se elimino con exito')
+        setTimeout(() => {
           
-        //   }
-        // })
-        // 
+          window.location.reload()
+         
+        },1000);
+ 
       }
-    )
+    })
+  
+  }
+
+  listaprendiz() {
+    throw new Error('Method not implemented.');
+    }
+    
+    
+
+
+  getAprendiz(document:number){ 
+    
+    const id =this.acrivaterouter.snapshot.params['id'];
+    this.serviceaprendiz.getAprendiz(document).subscribe({
+      next: data=>{
+        this.Aprendiz=data;
+        console.log(this.Aprendiz);
+        
+      },
+      error: err=>{
+        alert("usted no esta registrado en la base de datos")
+      }
+    })
   }
 
 }
